Fix ordering of cells
In the layout of the notebook one function was called before it got defined. Also removed some debugging lines from one of the function definitions, as well as fixed a bug in the end that only called one method when generating the playlist instead of all the four we defined.